feat: update app icons and Stream Deck plugin packaging
- Regenerate all app icons from new cc_icon.png with transparent corners
- Rename Stream Deck plugin to com.ccstatusbar.sdPlugin (reverse DNS)
- Add build.sh for Stream Deck plugin packaging
- Update release workflow to include Stream Deck plugin in DMG
- Add pr-check.yml for CI checks
