+++ Sources/CLI/FocusCommand.swift
@@ -1,0 +1,77 @@
+import ArgumentParser
+import Foundation
+
+struct FocusCommand: ParsableCommand {
+    static let configuration = CommandConfiguration(
+        commandName: "focus",
+        abstract: "Focus a terminal session"
+    )
+
+    @Option(name: .shortAndLong, help: "Session index (0-based)")
+    var index: Int?
+
+    @Option(name: .long, help: "Session ID (e.g., 'abc123:/dev/ttys001')")
+    var id: String?
+
+    func validate() throws {
+        if index == nil && id == nil {
+            throw ValidationError("Either --index or --id must be specified")
+        }
+        if index != nil && id != nil {
+            throw ValidationError("Cannot specify both --index and --id")
+        }
+    }
+
+    func run() {
+        let sessions = SessionStore.shared.getSessions()
+
+        guard !sessions.isEmpty else {
+            printError("No active sessions")
+            return
+        }
+
+        let session: Session?
+
+        if let index = index {
+            guard index >= 0 && index < sessions.count else {
+                printError("Index \(index) out of range (0-\(sessions.count - 1))")
+                return
+            }
+            session = sessions[index]
+        } else if let id = id {
+            session = sessions.first { $0.id == id }
+            if session == nil {
+                printError("Session not found: \(id)")
+                return
+            }
+        } else {
+            session = nil
+        }
+
+        guard let targetSession = session else {
+            printError("Failed to find session")
+            return
+        }
+
+        let result = FocusManager.shared.focus(session: targetSession)
+
+        switch result {
+        case .success:
+            // Also acknowledge the session (mark as seen)
+            SessionStore.shared.acknowledgeSession(sessionId: targetSession.sessionId, tty: targetSession.tty)
+            print("Focused: \(targetSession.projectName)")
+        case .partialSuccess(let reason):
+            // Still acknowledge on partial success
+            SessionStore.shared.acknowledgeSession(sessionId: targetSession.sessionId, tty: targetSession.tty)
+            print("Partial: \(targetSession.projectName) (\(reason))")
+        case .notFound(let hint):
+            printError("Not found: \(hint)")
+        case .notRunning:
+            printError("Terminal is not running")
+        }
+    }
+
+    private func printError(_ message: String) {
+        FileHandle.standardError.write(Data("Error: \(message)\n".utf8))
+    }
+}
