+++ StreamDeckPlugin/com.ccstatusbar.sdPlugin/.gitignore
