+++ StreamDeckPlugin/com.ccstatusbar.sdPlugin/bin/plugin.js
@@ -1,14 +1,34 @@
+"use strict";
 /**
  * CC Status Bar - Stream Deck Plugin (TypeScript Source)
  *
  * This is the TypeScript source file for the plugin.
  * Compile with: tsc src/plugin.ts --outDir bin --target ES2020 --module CommonJS
  */
-
-import WebSocket from 'ws';
-import { execSync, exec } from 'child_process';
-import path from 'path';
-
+var __importDefault = (this && this.__importDefault) || function (mod) {
+    return (mod && mod.__esModule) ? mod : { "default": mod };
+};
+Object.defineProperty(exports, "__esModule", { value: true });
+const ws_1 = __importDefault(require("ws"));
+const child_process_1 = require("child_process");
+const fs_1 = require("fs");
+const path_1 = __importDefault(require("path"));
+// Debug log to file
+const fs_2 = require("fs");
+const LOG_DIR = path_1.default.join(process.env.HOME || '', 'Library/Logs/CCStatusBar');
+const LOG_FILE = path_1.default.join(LOG_DIR, 'streamdeck-plugin.log');
+try {
+    (0, fs_2.mkdirSync)(LOG_DIR, { recursive: true });
+}
+catch { }
+function debugLog(msg) {
+    try {
+        (0, fs_1.appendFileSync)(LOG_FILE, `${new Date().toISOString()} ${msg}\n`);
+    }
+    catch (e) {
+        console.error('Log error:', e);
+    }
+}
 // Stream Deck SDK events
 const Events = {
     DID_RECEIVE_SETTINGS: 'didReceiveSettings',
@@ -26,8 +46,7 @@ const Events = {
     PROPERTY_INSPECTOR_DID_APPEAR: 'propertyInspectorDidAppear',
     PROPERTY_INSPECTOR_DID_DISAPPEAR: 'propertyInspectorDidDisappear',
     SEND_TO_PLUGIN: 'sendToPlugin'
-} as const;
-
+};
 // Action UUIDs
 const Actions = {
     SESSION: 'com.ccstatusbar.session',
@@ -36,160 +55,83 @@ const Actions = {
     DICTATION: 'com.ccstatusbar.dictation',
     ENTER: 'com.ccstatusbar.enter',
     ESCAPE: 'com.ccstatusbar.escape'
-} as const;
-
-// Device type to columns mapping
-const DeviceColumns: Record<number, number> = {
-    0: 5,  // Standard (15 buttons, 3x5)
-    1: 3,  // Mini (6 buttons, 2x3)
-    2: 8,  // XL (32 buttons, 4x8)
-    7: 4,  // Plus (8 buttons + touch, 2x4)
-    9: 4   // Neo (8 buttons + touch, 2x4)
 };
-
-// Device info storage
-interface DeviceInfo {
-    type: number;
-    columns: number;
-    rows: number;
-}
-const devices = new Map<string, DeviceInfo>();
-
-// Types
-interface Session {
-    id: string;
-    project: string;
-    status: 'running' | 'waiting_input' | 'stopped';
-    path: string;
-    waiting_reason?: 'permission_prompt' | 'stop' | 'unknown';
-    is_acknowledged?: boolean;
-}
-
-interface SessionListResponse {
-    sessions: Session[];
-    offset: number;
-    total: number;
-}
-
-interface ButtonInfo {
-    row: number;
-    col: number;
-    buttonIndex: number;
-    deviceId?: string;
-    sessionNumber?: string; // "auto" or "1"-"10"
-}
-
-interface StreamDeckMessage {
-    event: string;
-    action?: string;
-    context?: string;
-    device?: string;
-    deviceInfo?: {
-        type: number;
-        size: {
-            columns: number;
-            rows: number;
-        };
-    };
-    payload?: {
-        coordinates?: {
-            row: number;
-            column: number;
-        };
-        settings?: {
-            sessionNumber?: string;
-            [key: string]: unknown;
-        };
-    };
-}
-
+// Device type to columns mapping
+const DeviceColumns = {
+    0: 5, // Standard (15 buttons, 3x5)
+    1: 3, // Mini (6 buttons, 2x3)
+    2: 8, // XL (32 buttons, 4x8)
+    7: 4, // Plus (8 buttons + touch, 2x4)
+    9: 4 // Neo (8 buttons + touch, 2x4)
+};
+const devices = new Map();
 // Plugin state
-let websocket: WebSocket | null = null;
-let pluginUUID: string | null = null;
-const sessionButtons = new Map<string, ButtonInfo>();
+let websocket = null;
+let pluginUUID = null;
+const sessionButtons = new Map();
 let currentOffset = 0;
 let totalSessions = 0;
-let sessions: Session[] = [];
-let pollInterval: NodeJS.Timeout | null = null;
-
+let sessions = [];
+let pollInterval = null;
 // CCStatusBar CLI path
-const CLI_PATH = path.join(
-    process.env.HOME || '',
-    'Library/Application Support/CCStatusBar/bin/CCStatusBar'
-);
-
+const CLI_PATH = path_1.default.join(process.env.HOME || '', 'Library/Application Support/CCStatusBar/bin/CCStatusBar');
 /**
  * Connect to Stream Deck
  */
-function connectElgatoStreamDeckSocket(
-    port: string,
-    uuid: string,
-    registerEvent: string,
-    _info: string
-): void {
+function connectElgatoStreamDeckSocket(port, uuid, registerEvent, _info) {
     pluginUUID = uuid;
-
-    websocket = new WebSocket(`ws://127.0.0.1:${port}`);
-
+    websocket = new ws_1.default(`ws://127.0.0.1:${port}`);
     websocket.on('open', () => {
+        debugLog('WebSocket connected');
         send({ event: registerEvent, uuid });
         startPolling();
     });
-
-    websocket.on('message', (data: WebSocket.Data) => {
-        const message = JSON.parse(data.toString()) as StreamDeckMessage;
+    websocket.on('message', (data) => {
+        const message = JSON.parse(data.toString());
         handleMessage(message);
     });
-
-    websocket.on('error', (error: Error) => {
+    websocket.on('error', (error) => {
         console.error('WebSocket error:', error);
     });
-
     websocket.on('close', () => {
         stopPolling();
     });
 }
-
 /**
  * Send message to Stream Deck
  */
-function send(payload: Record<string, unknown>): void {
-    if (websocket && websocket.readyState === WebSocket.OPEN) {
+function send(payload) {
+    if (websocket && websocket.readyState === ws_1.default.OPEN) {
         websocket.send(JSON.stringify(payload));
     }
 }
-
 /**
  * Handle incoming messages from Stream Deck
  */
-function handleMessage(data: StreamDeckMessage): void {
+function handleMessage(data) {
+    debugLog(`handleMessage: event=${data.event}, action=${data.action || 'none'}`);
     const { event, action, context, device, deviceInfo, payload } = data;
-
     switch (event) {
         case Events.DEVICE_DID_CONNECT:
             if (device && deviceInfo) {
                 handleDeviceDidConnect(device, deviceInfo);
             }
             break;
-
         case Events.WILL_APPEAR:
             if (action && context && payload) {
                 handleWillAppear(action, context, payload, device);
             }
             break;
-
         case Events.WILL_DISAPPEAR:
             if (action && context) {
                 handleWillDisappear(action, context);
             }
             break;
-
         case Events.KEY_DOWN:
             if (action && context) {
                 handleKeyDown(action, context);
             }
             break;
-
         case Events.DID_RECEIVE_SETTINGS:
             if (action && context && payload) {
                 handleDidReceiveSettings(action, context, payload);
@@ -197,14 +139,10 @@ function handleMessage(data: StreamDeckMessage): void {
             break;
     }
 }
-
 /**
  * Handle device connection
  */
-function handleDeviceDidConnect(
-    deviceId: string,
-    deviceInfo: NonNullable<StreamDeckMessage['deviceInfo']>
-): void {
+function handleDeviceDidConnect(deviceId, deviceInfo) {
     const columns = DeviceColumns[deviceInfo.type] || deviceInfo.size.columns;
     devices.set(deviceId, {
         type: deviceInfo.type,
@@ -212,15 +150,10 @@ function handleDeviceDidConnect(
         rows: deviceInfo.size.rows
     });
 }
-
 /**
  * Handle settings received from Property Inspector
  */
-function handleDidReceiveSettings(
-    action: string,
-    context: string,
-    payload: NonNullable<StreamDeckMessage['payload']>
-): void {
+function handleDidReceiveSettings(action, context, payload) {
     if (action === Actions.SESSION) {
         const buttonInfo = sessionButtons.get(context);
         if (buttonInfo) {
@@ -229,28 +162,19 @@ function handleDidReceiveSettings(
         }
     }
 }
-
 /**
  * Handle button appearing
  */
-function handleWillAppear(
-    action: string,
-    context: string,
-    payload: NonNullable<StreamDeckMessage['payload']>,
-    deviceId?: string
-): void {
+function handleWillAppear(action, context, payload, deviceId) {
     if (action === Actions.SESSION && payload.coordinates) {
         const { row, column } = payload.coordinates;
-
         // Get device columns (default to 5 for Standard)
         const device = deviceId ? devices.get(deviceId) : undefined;
         const columns = device?.columns || 5;
         const buttonIndex = row * columns + column;
-
         // Get session number from settings
         const sessionNumber = payload.settings?.sessionNumber || 'auto';
-
-        const buttonInfo: ButtonInfo = {
+        const buttonInfo = {
             row,
             col: column,
             buttonIndex,
@@ -261,34 +185,36 @@ function handleWillAppear(
         updateSessionButton(context, buttonInfo);
     }
 }
-
 /**
  * Handle button disappearing
  */
-function handleWillDisappear(action: string, context: string): void {
+function handleWillDisappear(action, context) {
     if (action === Actions.SESSION) {
         sessionButtons.delete(context);
     }
 }
-
 /**
  * Handle key press
  */
-function handleKeyDown(action: string, context: string): void {
+function handleKeyDown(action, context) {
+    debugLog(`KeyDown: action=${action}`);
     switch (action) {
         case Actions.SESSION:
             handleSessionClick(context);
             break;
         case Actions.SCROLL_UP:
+            debugLog('Executing handleScrollUp');
             handleScrollUp();
             break;
         case Actions.SCROLL_DOWN:
+            debugLog('Executing handleScrollDown');
             handleScrollDown();
             break;
         case Actions.DICTATION:
             handleDictation();
             break;
         case Actions.ENTER:
+            debugLog('Executing handleEnter');
             handleEnter();
             break;
         case Actions.ESCAPE:
@@ -296,216 +222,176 @@ function handleKeyDown(action: string, context: string): void {
             break;
     }
 }
-
 /**
  * Handle escape key
  */
-function handleEscape(): void {
-    const script = `
-        tell application "System Events"
-            key code 53
-        end tell
-    `;
-
-    exec(`osascript -e '${script}'`, (error) => {
+function handleEscape() {
+    (0, child_process_1.exec)(`osascript -e 'tell application "System Events" to key code 53'`, (error) => {
         if (error) {
             console.error('Escape key error:', error.message);
         }
     });
 }
-
 /**
  * Handle session button click
  */
-function handleSessionClick(context: string): void {
+function handleSessionClick(context) {
     const buttonInfo = sessionButtons.get(context);
-    if (!buttonInfo) return;
-
+    if (!buttonInfo)
+        return;
     // Determine session index based on settings
-    let sessionIndex: number;
+    let sessionIndex;
     if (buttonInfo.sessionNumber && buttonInfo.sessionNumber !== 'auto') {
         sessionIndex = parseInt(buttonInfo.sessionNumber, 10) - 1;
-    } else {
+    }
+    else {
         sessionIndex = currentOffset + buttonInfo.buttonIndex;
     }
-
-    if (sessionIndex < 0 || sessionIndex >= sessions.length) return;
-
+    if (sessionIndex < 0 || sessionIndex >= sessions.length)
+        return;
     try {
-        execSync(`"${CLI_PATH}" focus --index ${sessionIndex}`, {
+        (0, child_process_1.execSync)(`"${CLI_PATH}" focus --index ${sessionIndex}`, {
             encoding: 'utf-8',
             timeout: 5000
         });
-    } catch (error) {
-        console.error('Focus error:', (error as Error).message);
+    }
+    catch (error) {
+        console.error('Focus error:', error.message);
     }
 }
-
 /**
- * Handle scroll up
+ * Handle up arrow key
  */
-function handleScrollUp(): void {
-    if (currentOffset > 0) {
-        currentOffset = Math.max(0, currentOffset - 10);
-        updateAllSessionButtons();
-    }
+function handleScrollUp() {
+    (0, child_process_1.exec)(`osascript -e 'tell application "System Events" to key code 126'`, (error) => {
+        if (error) {
+            console.error('Up arrow key error:', error.message);
+        }
+    });
 }
-
 /**
- * Handle scroll down
+ * Handle down arrow key
  */
-function handleScrollDown(): void {
-    if (currentOffset + 10 < totalSessions) {
-        currentOffset += 10;
-        updateAllSessionButtons();
-    }
+function handleScrollDown() {
+    (0, child_process_1.exec)(`osascript -e 'tell application "System Events" to key code 125'`, (error) => {
+        if (error) {
+            console.error('Down arrow key error:', error.message);
+        }
+    });
 }
-
 /**
- * Handle dictation toggle via Edit menu
+ * Handle dictation toggle via CCStatusBar CLI
  */
-function handleDictation(): void {
-    const script = `
-        tell application "System Events"
-            set p to first process whose frontmost is true
-            try
-                tell menu bar item "編集" of menu bar 1 of p
-                    click menu item "音声入力を開始" of menu 1
-                end tell
-            on error
-                try
-                    tell menu bar item "Edit" of menu bar 1 of p
-                        click menu item "Start Dictation" of menu 1
-                    end tell
-                end try
-            end try
-        end tell
-    `;
-
-    exec(`osascript -e '${script}'`, (error) => {
+function handleDictation() {
+    // Use CCStatusBar CLI which handles CGEvent double-Fn tap with AppleScript fallback
+    (0, child_process_1.exec)(`"${CLI_PATH}" dictation`, (error) => {
         if (error) {
             console.error('Dictation error:', error.message);
         }
     });
 }
-
 /**
  * Handle enter key
  */
-function handleEnter(): void {
-    const script = `
-        tell application "System Events"
-            key code 36
-        end tell
-    `;
-
-    exec(`osascript -e '${script}'`, (error) => {
+function handleEnter() {
+    (0, child_process_1.exec)(`osascript -e 'tell application "System Events" to key code 36'`, (error) => {
         if (error) {
             console.error('Enter key error:', error.message);
         }
     });
 }
-
 /**
  * Start polling for session updates
  */
-function startPolling(): void {
+function startPolling() {
     fetchSessions();
     pollInterval = setInterval(fetchSessions, 1000);
 }
-
 /**
  * Stop polling
  */
-function stopPolling(): void {
+function stopPolling() {
     if (pollInterval) {
         clearInterval(pollInterval);
         pollInterval = null;
     }
 }
-
 /**
  * Fetch sessions from CLI
  */
-function fetchSessions(): void {
+function fetchSessions() {
     try {
-        const output = execSync(`"${CLI_PATH}" list --json`, {
+        const output = (0, child_process_1.execSync)(`"${CLI_PATH}" list --json`, {
             encoding: 'utf-8',
             timeout: 5000
         });
-
-        const data = JSON.parse(output) as SessionListResponse;
+        const data = JSON.parse(output);
         sessions = data.sessions || [];
         totalSessions = data.total || 0;
-
         updateAllSessionButtons();
-    } catch {
+    }
+    catch {
         sessions = [];
         totalSessions = 0;
         updateAllSessionButtons();
     }
 }
-
 /**
  * Update all session buttons
  */
-function updateAllSessionButtons(): void {
+function updateAllSessionButtons() {
     for (const [context, buttonInfo] of sessionButtons) {
         updateSessionButton(context, buttonInfo);
     }
 }
-
 /**
  * Update a single session button
  */
-function updateSessionButton(context: string, buttonInfo: ButtonInfo): void {
+function updateSessionButton(context, buttonInfo) {
     // Determine session index based on settings
-    let sessionIndex: number;
+    let sessionIndex;
     if (buttonInfo.sessionNumber && buttonInfo.sessionNumber !== 'auto') {
         // Fixed session number (1-based, convert to 0-based)
         sessionIndex = parseInt(buttonInfo.sessionNumber, 10) - 1;
-    } else {
+    }
+    else {
         // Auto: use button position with current offset
         sessionIndex = currentOffset + buttonInfo.buttonIndex;
     }
-
     const session = sessions[sessionIndex];
-
     if (!session) {
         setImage(context, createEmptyButtonSVG());
         return;
     }
-
     // Determine background color based on status and acknowledged state
     // Match menu bar behavior: acknowledged waiting_input shows as green
-    let bgColor: string;
+    let bgColor;
     if (session.is_acknowledged) {
-        bgColor = '#34C759';  // Green (acknowledged)
-    } else if (session.status === 'running') {
-        bgColor = '#34C759';  // Green
-    } else if (session.status === 'waiting_input') {
-        bgColor = session.waiting_reason === 'permission_prompt' ? '#FF3B30' : '#FFCC00';
-    } else {
-        bgColor = '#8E8E93';  // Gray
+        bgColor = '#34C759'; // Green (acknowledged)
     }
-
+    else if (session.status === 'running') {
+        bgColor = '#34C759'; // Green
+    }
+    else if (session.status === 'waiting_input') {
+        bgColor = session.waiting_reason === 'permission_prompt' ? '#FF3B30' : '#FFCC00';
+    }
+    else {
+        bgColor = '#8E8E93'; // Gray
+    }
     const svg = createSessionButtonSVG(session.project, bgColor, buttonInfo.buttonIndex);
     setImage(context, svg);
 }
-
 /**
  * Create SVG for session button (72x72, max 3 lines, vertically centered)
  */
-function createSessionButtonSVG(projectName: string, bgColor: string, buttonIndex: number = 0): string {
+function createSessionButtonSVG(projectName, bgColor, buttonIndex = 0) {
     const textColor = bgColor === '#FFCC00' ? '#000000' : '#FFFFFF';
     const maxCharsPerLine = 6;
     const fontSize = 18;
     const lineHeight = 20;
-
     // Split into lines (max 3)
-    const lines: string[] = [];
+    const lines = [];
     let remaining = projectName;
-
     while (remaining.length > 0 && lines.length < 3) {
         if (remaining.length <= maxCharsPerLine) {
             lines.push(remaining);
@@ -514,55 +400,45 @@ function createSessionButtonSVG(projectName: string, bgColor: string, buttonInde
         // Try to split at hyphen/underscore within range
         const chunk = remaining.substring(0, maxCharsPerLine + 2);
         const splitMatch = chunk.match(/[-_]/);
-        let splitIndex: number;
+        let splitIndex;
         if (splitMatch && splitMatch.index !== undefined && splitMatch.index > 0 && splitMatch.index <= maxCharsPerLine) {
             splitIndex = splitMatch.index + 1;
-        } else {
+        }
+        else {
             splitIndex = maxCharsPerLine;
         }
         lines.push(remaining.substring(0, splitIndex));
         remaining = remaining.substring(splitIndex);
     }
-
     // Truncate last line if there's more text
     if (remaining.length > 0 && lines.length === 3) {
         lines[2] = lines[2].substring(0, maxCharsPerLine - 1) + '…';
     }
-
     const sessionNum = currentOffset + buttonIndex + 1;
-
     // Calculate vertical center (36 is center of 72px canvas)
     const totalTextHeight = lines.length * lineHeight;
     const startY = 36 - (totalTextHeight / 2) + (lineHeight / 2) + 4; // +4 for session number offset
-
-    const textElements = lines.map((line, i) =>
-        `<text x="36" y="${startY + i * lineHeight}" font-family="system-ui, -apple-system, sans-serif" font-size="${fontSize}" font-weight="bold" fill="${textColor}" text-anchor="middle" dominant-baseline="middle">${escapeXml(line)}</text>`
-    ).join('\n');
-
+    const textElements = lines.map((line, i) => `<text x="36" y="${startY + i * lineHeight}" font-family="system-ui, -apple-system, sans-serif" font-size="${fontSize}" font-weight="bold" fill="${textColor}" text-anchor="middle" dominant-baseline="middle">${escapeXml(line)}</text>`).join('\n');
     const svg = `<svg xmlns="http://www.w3.org/2000/svg" width="72" height="72" viewBox="0 0 72 72">
 <rect width="72" height="72" rx="8" fill="${bgColor}"/>
 <text x="4" y="12" font-family="system-ui, -apple-system, sans-serif" font-size="10" font-weight="bold" fill="${textColor}" opacity="0.7">${sessionNum}</text>
 ${textElements}
 </svg>`;
-
     return `data:image/svg+xml,${encodeURIComponent(svg)}`;
 }
-
 /**
  * Create SVG for empty button (72x72)
  */
-function createEmptyButtonSVG(): string {
+function createEmptyButtonSVG() {
     const svg = `<svg xmlns="http://www.w3.org/2000/svg" width="72" height="72" viewBox="0 0 72 72">
 <rect width="72" height="72" rx="8" fill="#2C2C2E"/>
 </svg>`;
-
     return `data:image/svg+xml,${encodeURIComponent(svg)}`;
 }
-
 /**
  * Escape XML special characters
  */
-function escapeXml(str: string): string {
+function escapeXml(str) {
     return str
         .replace(/&/g, '&amp;')
         .replace(/</g, '&lt;')
@@ -570,11 +446,10 @@ function escapeXml(str: string): string {
         .replace(/"/g, '&quot;')
         .replace(/'/g, '&apos;');
 }
-
 /**
  * Set button image
  */
-function setImage(context: string, imageData: string): void {
+function setImage(context, imageData) {
     send({
         event: 'setImage',
         context,
@@ -584,14 +459,12 @@ function setImage(context: string, imageData: string): void {
         }
     });
 }
-
 // Parse command line arguments and connect
 const args = process.argv.slice(2);
-let port: string | undefined;
-let uuid: string | undefined;
-let registerEvent: string | undefined;
-let info: string | undefined;
-
+let port;
+let uuid;
+let registerEvent;
+let info;
 for (let i = 0; i < args.length; i++) {
     switch (args[i]) {
         case '-port':
@@ -608,7 +481,6 @@ for (let i = 0; i < args.length; i++) {
             break;
     }
 }
-
 if (port && uuid && registerEvent) {
     connectElgatoStreamDeckSocket(port, uuid, registerEvent, info || '');
 }
