+++ StreamDeckPlugin/com.ccstatusbar.sdPlugin/package.json
