+++ StreamDeckPlugin/com.ccstatusbar.sdPlugin/profiles/README.md
