+++ StreamDeckPlugin/com.ccstatusbar.sdPlugin/tsconfig.json
