+++ StreamDeckPlugin/com.ccstatusbar.sdPlugin/ui/session-settings.html
