+++ StreamDeckPlugin/com.ccstatusbar.sdPlugin/ui/settings.html
